Modifications galore

+++ Resources/ItemEndpoints.py
@@ -1,6 +1,12 @@
-import sqlite3
+
 from flask_restful import Resource,reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+                                jwt_required,
+                                get_jwt_claims,
+                                jwt_optional,
+                                get_jwt_identity,
+                                fresh_jwt_required
+                                )
 
 from Models.itemModels import ItemModel
 
@@ -26,7 +32,7 @@ class Item(Resource):
             return item.json()
         return {"message":"Item not found"},404
 
-    @jwt_required()
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {"message":"An item with name '{}'already exists.".format(name)},400
@@ -39,14 +45,17 @@ class Item(Resource):
             return {"message":"An Error occurred"},500
         return item.json(),201
     
-    @jwt_required()
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return{"message":"Admin priviledge required"},401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
         return {"message":"item {} deleted".format(name)}
 
-    @jwt_required()
+    @jwt_required
     def put(self, name):
         data = Item.parser.parse_args()
 
@@ -61,8 +70,16 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return{ "items":[item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return{ "items":items},200
+        return{
+            "items":[item["name"] for item in items],
+            "message":"More data available if you log in"
+        }
         
-        # Use list comprehension or lambda
+        # Use list comprehension or lambda function
         # return {"items":list(map(lambda x:x.json(),ItemModel.query.all()))}
